requirements.txt
diffusers==0.10.2
einops==0.4.1
fire==0.4.0
matplotlib==3.5.1
numpy>=1.23.3
Pillow==9.4.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.63.1
transformers>=4.19.2
